Removed `vitestSetup.ts` from vitest setup files
