+++ packages/attribute-registry-process/test/readModelService.test.ts
@@ -15,7 +15,7 @@ import {
   addOneAttribute,
   attributeRegistryService,
   readModelService,
-} from "./vitestSetup.js";
+} from "./utils.js";
 
 const mockAttribute = getMockAttribute();
