+++ packages/attribute-registry-process/vitest.config.ts
@@ -2,7 +2,6 @@ import { defineConfig } from "vitest/config";
 
 export default defineConfig({
   test: {
-    setupFiles: ["./test/vitestSetup.ts"],
     globalSetup: ["./test/vitestGlobalSetup.ts"],
     testTimeout: 60000,
     hookTimeout: 60000,
